feat: split page with file base component and add footer
