+++ app/types.ts
@@ -1,0 +1,6 @@
+export type CardType = 'Weibo' | 'Zhihu' | 'Netease'
+
+export interface Word {
+  url: string
+  title: string
+}
